Измненеие архитектуры вновь
