+++ email/18-ISbo-2a/configs/config.py
@@ -1,0 +1,14 @@
+"""
+Ссылка на google.drive c инструкцией
+>>> https://docs.google.com/document/d/1rARH0Hdzq7w9ygDbpsrkZ9Pnw82p6WSHQ5AeUT8jOzA/edit
+"""
+
+# Id таблицы
+SPREAD_SHEET_ID = '1gOX8T8ihy3J1khhC16U1qDwaI-K6ndkp9LFWAHncuWA'
+SPREAD_SHEET_ID_INIT = '1iYcDfGAbOhe9K2z1OTl1_0im3IPGkqxnv33LGccgEb4'
+# Ключ Google API для работы с таблицей
+CREDENTIALS_FILE = 'Example.json'
+# Ключ Google API для получения доступа к аккаунту
+CREDENTIALS_FILE_SERVICE = 'credentials.json'
+# Email для тестирования
+TESTING_EMAIL = '0sashasmirnov0@gmail.com'
