+++ email/18-ISbo-2a/utils/crypto.py
@@ -1,8 +1,10 @@
-from cryptography.fernet import Fernet
 import os
+from cryptography.fernet import Fernet
+#from . import log_method
 import log_method
 
 cripher = Fernet(b'D7Zh4RcKcxIRDOrxB68u98_WT_31qlMhXwFTMByCAYc=')
+
 
 @log_method.log_method_info
 def crypt_file(filename):
@@ -41,10 +43,11 @@ def decrypt_file(filename):
 
 
 if __name__ == '__main__':
-    crypt_file('credentials.json')
-    crypt_file('Example.json')
-    crypt_file('config.py')
-    crypt_file('token.pickle')
-    #decrypt_file('Example.json.bin')
-    #decrypt_file('credentials.json.bin')
-    #decrypt_file('config.py.bin')
+    # crypt_file(r'../configs/credentials.json')
+    # crypt_file(r'../configs/Example.json')
+    # crypt_file(r'../configs/config.py')
+    # crypt_file(r'../configs/token.pickle')
+    decrypt_file(r'../configs/credentials.json.bin')
+    # decrypt_file(r'../configs/Example.json.bin')
+    # decrypt_file(r'../configs/config.py.bin')
+    # decrypt_file(r'../configs/token.pickle.bin')
